pyproject.toml
[project]
name = "huggingface-datasets-with-tigris"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets[s3]>=3.6.0",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.1",
    "s3fs>=2025.3.0",
]
